+++ rteam_prozorro/wizards/__init__.py
@@ -1,0 +1,1 @@
+from . import prozorro_subscription_test